make it possible to use testkit with ScalaTest 3.0.x, 3.1.x & 3.2.x
* Deprecate existing *RuleSuite as they cause runtime failure when
  used in a classpath where ScalaTest 3.0.x gets evicted by 3.2.x
* testkit clients must now mix-in themselves an implementation of
  TestRegistration (as this trait moved between 3.0.x & 3.2.x)
  like [Any]FunSuiteLike
* Run https://github.com/scalatest/autofix/tree/master/3.1.x on unit
