+++ .scala-steward.conf
@@ -1,1 +1,5 @@
 pullRequests.frequency = "0 0 1,15 * ?"
+updates.pin = [
+  # don't bump, to avoid forcing breaking changes on clients via eviction
+  { groupId = "org.scalatest", artifactId = "scalatest", version = "3.0.8" },
+]
