+++ scalafix-testkit/src/main/scala/scalafix/testkit/SemanticRuleSuite.scala
@@ -1,28 +1,23 @@
 package scalafix.testkit
 
-import java.nio.charset.StandardCharsets
-import org.scalatest.BeforeAndAfterAll
-import org.scalatest.FunSuite
-import org.scalatest.exceptions.TestFailedException
 import scala.meta._
-import scala.meta.internal.io.FileIO
-import scalafix.internal.reflect.ClasspathOps
+import org.scalatest.FunSpecLike
 import scalafix.internal.reflect.RuleCompiler
-import scalafix.internal.testkit.AssertDiff
-import scalafix.internal.testkit.CommentAssertion
-import scalafix.internal.testkit.EndOfLineAssertExtractor
-import scalafix.internal.testkit.MultiLineAssertExtractor
+import scalafix.internal.testkit.{
+  EndOfLineAssertExtractor,
+  MultiLineAssertExtractor
+}
 import scalafix.v0.SemanticdbIndex
-import java.nio.file.Files
 
-/** Construct a test suite for running semantic Scalafix rules. */
-abstract class SemanticRuleSuite(
-    val props: TestkitProperties,
-    val isSaveExpect: Boolean
-) extends FunSuite
-    with DiffAssertions
-    with BeforeAndAfterAll { self =>
-
+@deprecated(
+  "Use AbstractSemanticRuleSuite with the styling trait of your choice mixed-in (*SpecLike or *SuiteLike)",
+  "0.9.18"
+)
+class SemanticRuleSuite(
+    override val props: TestkitProperties,
+    override val isSaveExpect: Boolean
+) extends AbstractSemanticRuleSuite
+    with FunSpecLike {
   def this(props: TestkitProperties) = this(props, isSaveExpect = false)
   def this() = this(TestkitProperties.loadFromResources())
 
@@ -35,78 +30,6 @@ abstract class SemanticRuleSuite(
       inputSourceroot: AbsolutePath,
       expectedOutputSourceroot: Seq[AbsolutePath]
   ) = this()
-
-  private def scalaVersion: String = scala.util.Properties.versionNumberString
-  private def scalaVersionDirectory: Option[String] =
-    if (scalaVersion.startsWith("2.11")) Some("scala-2.11")
-    else if (scalaVersion.startsWith("2.12")) Some("scala-2.12")
-    else if (scalaVersion.startsWith("2.13")) Some("scala-2.13")
-    else None
-
-  def evaluateTestBody(diffTest: RuleTest): Unit = {
-    val (rule, sdoc) = diffTest.run.apply()
-    rule.beforeStart()
-    val (fixed, messages) =
-      try rule.semanticPatch(sdoc, suppress = false)
-      finally rule.afterComplete()
-    val tokens = fixed.tokenize.get
-    val obtained = SemanticRuleSuite.stripTestkitComments(tokens)
-    val expected = diffTest.path.resolveOutput(props) match {
-      case Right(file) =>
-        FileIO.slurp(file, StandardCharsets.UTF_8)
-      case Left(err) =>
-        if (fixed == sdoc.input.text) {
-          // rule is a linter, no need for an output file.
-          obtained
-        } else {
-          fail(err)
-        }
-    }
-
-    val expectedLintMessages = CommentAssertion.extract(sdoc.tokens)
-    val diff = AssertDiff(messages, expectedLintMessages)
-
-    if (diff.isFailure) {
-      println("###########> Lint       <###########")
-      println(diff.toString)
-    }
-
-    val result = compareContents(obtained, expected)
-    if (result.nonEmpty) {
-      println("###########> Diff       <###########")
-      println(error2message(obtained, expected))
-    }
-
-    val isTestFailure = result.nonEmpty || diff.isFailure
-    diffTest.path.resolveOutput(props) match {
-      case Right(output) if isTestFailure && isSaveExpect =>
-        println(s"promoted expect test: $output")
-        Files.write(output.toNIO, obtained.getBytes(StandardCharsets.UTF_8))
-      case _ =>
-    }
-
-    if (isTestFailure) {
-      throw new TestFailedException("see above", 0)
-    }
-  }
-
-  def runOn(diffTest: RuleTest): Unit = {
-    test(diffTest.path.testName) {
-      evaluateTestBody(diffTest)
-    }
-  }
-
-  lazy val testsToRun = {
-    val symtab = ClasspathOps.newSymbolTable(props.inputClasspath)
-    val classLoader = ClasspathOps.toClassLoader(props.inputClasspath)
-    val tests = TestkitPath.fromProperties(props)
-    tests.map { test =>
-      RuleTest.fromPath(props, test, classLoader, symtab)
-    }
-  }
-  def runAllTests(): Unit = {
-    testsToRun.foreach(runOn)
-  }
 }
 
 object SemanticRuleSuite {
