+++ scalafix-tests/unit/src/test/scala/scalafix/tests/cli/AutoClasspathSuite.scala
@@ -3,10 +3,10 @@ package scalafix.tests.cli
 import java.io.File
 import scala.meta.io.AbsolutePath
 import scala.meta.io.Classpath
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 import scalafix.internal.reflect.ClasspathOps
 
-class AutoClasspathSuite extends FunSuite {
+class AutoClasspathSuite extends AnyFunSuite {
   test("--classpath=auto") {
     val tmp = File.createTempFile("foo", "bar")
     assert(tmp.delete())
