+++ scalafix-tests/unit/src/test/scala/scalafix/tests/cli/CliGitDiffSuite.scala
@@ -5,13 +5,13 @@ import java.io.PrintStream
 import java.nio.charset.StandardCharsets
 import java.nio.file.Path
 import org.scalatest._
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 import scalafix.cli.ExitStatus
 import scalafix.testkit.DiffAssertions
 import scalafix.internal.tests.utils.{Fs, Git}
 import scalafix.internal.tests.utils.SkipWindows
 
-class CliGitDiffSuite extends FunSuite with DiffAssertions {
+class CliGitDiffSuite extends AnyFunSuite with DiffAssertions {
   gitTest("addition", SkipWindows) { (fs, git, cli) =>
     val oldCode = "old.scala"
     val newCode = "new.scala"
