+++ scalafix-tests/unit/src/test/scala/scalafix/tests/config/ExplicitResultTypesConfigSuite.scala
@@ -1,11 +1,11 @@
 package scalafix.tests.config
 
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 import scalafix.internal.rule._
 import scalafix.v1.Configuration
 import scalafix.internal.reflect.ClasspathOps
 
-class ExplicitResultTypesConfigSuite extends FunSuite {
+class ExplicitResultTypesConfigSuite extends AnyFunSuite {
   test("Unsupported Scala version") {
     val scalaVersion = "2.12.0"
     val classpath = ClasspathOps.thisClasspath.entries
