+++ scalafix-tests/unit/src/test/scala/scalafix/tests/reflect/RuleDecoderSuite.scala
@@ -7,10 +7,10 @@ import metaconfig.Conf
 import metaconfig.ConfDecoder
 import scala.meta.io.AbsolutePath
 import scala.meta.io.RelativePath
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 import scalafix.v1.RuleDecoder
 
-class RuleDecoderSuite extends FunSuite {
+class RuleDecoderSuite extends AnyFunSuite {
   val cwd: AbsolutePath = AbsolutePath(BuildInfo.baseDirectory)
     .resolve("scalafix-tests")
     .resolve("unit")
