+++ scalafix-tests/unit/src/test/scala/scalafix/tests/util/ScalametaStructureSuite.scala
@@ -1,11 +1,11 @@
 package scalafix.tests.util
 
 import scala.meta._
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 import scalafix.testkit.DiffAssertions
 import scalafix.v1._
 
-class ScalametaStructureSuite extends FunSuite with DiffAssertions {
+class ScalametaStructureSuite extends AnyFunSuite with DiffAssertions {
   test("pretty(t)") {
     val obtained = q"a.b.c.d".structureWidth(1)
     val expected =
